Use a more flexible package removal logic
We now filter the packages by what's installed before trying to remove
them. This is all so that we don't fail the workflow at the cleanup
stage in case something in the list is not actually present (apt-get
remove/purge will error out if the package you're trying to remove is
not present and there's no flag to make it ignore those).