Separate available version from built version
Update logic to always save the latest version from Proxmox but
check against the latest built version to determine whether to
trigger a build.

Fixed patching for kernel version 5.15 so that the relaxablermrr
suffix is correctly applied.

Fixed artifact uploading for newer 6.2.16 kernels which started
failing due to internal renaming (pve-kernel-* to
proxmox-kernel-*).